==> _previous/clustering/sjvin/cluster_baselineTest/02-5_preparation_donnees_commerce.ipynb <==
# Préparation des variables commerce

library(tidyverse)
library(factoextra)

# Charger les données
df_pilot1_2022 <- readRDS("../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/01_pilot1_2022.rds")

# Charger le fichier des variables
source("01-1_variables.R")

df_commerce <- df_pilot1_2022 %>%
  select(all_of(variables_commerce))

# Boucle pour afficher la table pour chaque variable
for (v in variables_commerce) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_commerce[[v]]))
    cat("\n\n")
  }
}

### Transformations

- cons_brand : Aucun changement

- cons_coffee : Regrouper SC (second cup) avec Other

# Regroupement de cons_coffee
#-----------------------------

df_commerce <- df_commerce %>%

  mutate(
    cons_coffee_Other = if_else(
      cons_coffee_SC == 1,
      1,
      cons_coffee_Other
    )
  ) %>%
  select(-cons_coffee_SC)

variables_commerce_clust <- c(
  "id",
  "cons_brand_MaR",
  "cons_brand_BInd",
  "cons_brand_ChainesB",
  "cons_brand_GSurf",
  "cons_brand_OnlineOnly",
  "cons_brand_Frip",
  "cons_brand_Other",
  "cons_coffee_TimH",
  "cons_coffee_Starbucks",
  "cons_coffee_McDo",
  "cons_coffee_Other",
  "cons_coffee_place_ind"
)

# Boucle pour afficher la table pour chaque variable
for (v in variables_commerce_clust) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_commerce[[v]]))
    cat("\n\n")
  }
}

# Sauvegarder les données préparées
saveRDS(df_commerce, file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/qc2022/02_pilot1_2022_commerce.rds")

---

# SUGGESTION : Activité de Clustering

Si souhaité, effectuer un exercice de clustering avec ce sous-groupe de variables pour mieux saisir les données et leur relation. 

Cet exercice est purement exploratoire et sert à voir si les données peuvent être utilisées pour faire du clustering.
